+++ django_api_forms/fields.py
@@ -1,10 +1,8 @@
 import typing
-
 from enum import Enum
 
 from django.core.exceptions import ValidationError
 from django.forms import Field
-
 from django.utils.translation import gettext_lazy as _
 
 from .exceptions import RequestValidationError
@@ -82,7 +80,7 @@ class FormField(Field):
 
     def to_python(self, value) -> typing.Union[typing.Dict, None]:
         if not value:
-            return None
+            return {}
 
         form = self._form(value)
         if form.is_valid():
@@ -125,8 +123,10 @@ class EnumField(Field):
     }
 
     def __init__(self, enum: typing.Type, **kwargs):
-        if not issubclass(enum, Enum):
-            raise RuntimeError("Invalid Field type passed into FieldList!")
+        # isinstance(enum, type) prevents "TypeError: issubclass() arg 1 must be a class"
+        # based on: https://github.com/samuelcolvin/pydantic/blob/v0.32.x/pydantic/utils.py#L260-L261
+        if not (isinstance(enum, type) and issubclass(enum, Enum)):
+            raise RuntimeError("Invalid Enum type passed into EnumField!")
 
         self.enum = enum
 
@@ -137,7 +137,7 @@ class EnumField(Field):
             try:
                 return self.enum(value)
             except ValueError:
-                raise ValidationError(f"Invalid enum value {value} passed to {type(self.enum)}")
+                raise ValidationError(f"Invalid enum value {value} passed to {self.enum}")
         return None